Implement approval voting component
